+++ src/FamilyHubs.ServiceDirectory.Admin.Core/Models/JourneyFlow.cs
@@ -16,26 +16,3 @@ public enum JourneyFlow
     /// </summary>
     Edit
 }
-
-//todo: move to web, or keep together?
-public static class JourneyFlowExtensions
-{
-    public static string ToUrlString(this JourneyFlow flow)
-    {
-        return flow.ToString().ToLowerInvariant();
-    }
-
-    public static JourneyFlow FromUrlString(string? urlString)
-    {
-        //todo: have a default for when flow is not specified?
-        ArgumentNullException.ThrowIfNullOrEmpty(urlString);
-
-        if (!Enum.TryParse(urlString, true, out JourneyFlow flow))
-        {
-            //todo: throw here, or let consumer handle it?
-            throw new InvalidOperationException($"Invalid {nameof(JourneyFlow)} string representation: {urlString}");
-        }
-
-        return flow;
-    }
-}
